requirements.txt
pygame>=2.5.0
opencv-python>=4.8.0
pillow>=10.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0